feat(test): implement lemniscate test bench and remove obsolete files
