+++ perception/autoware_multi_object_tracker/CMakeLists.txt
@@ -85,7 +85,6 @@ if(BUILD_TESTING)
     test/test_bench.cpp
     test/test_utils.cpp
     test/test_bench_association.cpp
-    test/test_bench_association_lemniscate.cpp
   )
   add_definitions(-D_SRC_RESOURCES_DIR_PATH="${PROJECT_SOURCE_DIR}/test/data/")
   ament_add_ros_isolated_gtest(test_multi_object_tracker ${test_files})
